requirements-dev.txt
flake8==7.3.0
pytest==8.4.1
pytest-asyncio==0.23.6
requests==2.32.4  # HTTP client used by update scripts
beautifulsoup4==4.13.4  # HTML parsing for data scraping
black==25.1.0
isort==6.0.1
mypy==1.16.1
pre-commit==4.2.0
filelock==3.18.0
prometheus_client==0.22.1  # Metrics exposure for FastAPI app
numpy==2.3.1
structlog==25.4.0
fastapi==0.116.0
uvicorn==0.35.0  # ASGI server for development
sentence-transformers==5.0.0  # Embedding models for retrieval
faiss-cpu==1.11.0  # Optional fast vector search backend
tiktoken==0.9.0  # Token counting utility for LLM usage
openai==1.98.0  # OpenAI API client
httpx==0.27.2
httpx_ws==0.7.2
starlette==0.46.2
pydantic>=2.7,<3.0
xmlschema==4.1.0
PyYAML==6.0.1
psutil==7.0.0
defusedxml==0.7.1  # Secure XML parsing to prevent XXE attacks

# JWT support
PyJWT==2.8.0

# tracing
opentelemetry-api==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-exporter-jaeger==1.21.0

# password hashing
bcrypt==4.3.0
redis==5.2.1  # Redis async client for session management

dvc[s3]==3.61.0  # Dataset versioning tool with S3 remote
playwright==1.43.0  # UI end-to-end testing
